requirements.txt
langchain
sentence-transformers
langchain_anthropic
langchain-community
jq
langchain-huggingface
faiss-cpu
flask
numpy